requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==3.1.12
djangorestframework==3.11.0
djongo==1.3.7
executing==2.2.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
kombu==5.5.4
legacy-cgi==2.6.4
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymongo==3.11.4
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.2.4
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
